+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.2.2 2012-09-17
+
+* Do not show settings link on plugins.php when WooCommerce is inactive.
+
+## 1.2.1 2021-09-13
+
+* When no validator is available for the address, do not mark the order "bad address"! Just add an order note.
+
+
 ## 1.2.0 2021-08-20
 
 * Updated address now saves to customer user account